Changes to the `SLAViolated` chaincode function.
The function now takes care of setting the right status and transferring the
ownership of the value to the right stakeholder.
To do that we removed these info from the Violation data the producer creates.

TODO: The SLA struct is different inside the chaincode and outside, check how this affects performance.